requirements.txt
alembic==1.13.2
blinker==1.9.0
cffi==1.17.1
chardet==5.2.0
click==8.1.7
cryptography==45.0.6
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==22.0.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.2
packaging==25.0
pandas==2.2.2
pillow==11.3.0
pluggy==1.6.0
pre-commit==4.3.0
pycparser==2.22
Pygments==2.19.2
pypdf==4.3.1
pytest==8.4.1
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
RapidFuzz==3.13.0
reportlab==4.2.5
ruff==0.12.5
six==1.17.0
SQLAlchemy==2.0.31
thefuzz==0.22.1
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.0.3
WTForms==3.1.2
email-validator==2.2.0
python-dotenv==1.0.1
PyYAML==6.0.2